feat: реализована работа приложения с БД H2, добавлены жанры, рейтинг МПА для фильмов, добавлены интеграционные тесты для ДАО

+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -2,13 +2,14 @@ package ru.yandex.practicum.filmorate.controller;
 
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.service.UserService;
+import ru.yandex.practicum.filmorate.util.ValidatorControllers;
 
 import java.util.*;
 
-import org.springframework.web.bind.annotation.*;
-import org.springframework.validation.annotation.Validated;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.validation.annotation.Validated;
 
 import javax.validation.Valid;
 
@@ -18,24 +19,30 @@ import javax.validation.Valid;
 @Validated
 @RequiredArgsConstructor
 public class UserController {
+
     private final UserService userService;
 
     @PostMapping
     @Validated
     // создание пользователя
     public User create(@Valid @RequestBody User user) {
-        userService.create(user);
-        log.debug("Добавлен новый пользователь: {}", user);
-        return user;
+        user = ValidatorControllers.validateUser(user);
+        User newUser = userService.create(user).get();
+        log.debug("Добавлен новый пользователь: {}", newUser);
+        return newUser;
     }
 
     @PutMapping
     @Validated
     // обновление пользователя
     public User update(@Valid @RequestBody User user) {
-        userService.update(user);
-        log.debug("Обновлен пользователь: {}", user);
-        return user;
+        if (findUserById(user.getId()).isEmpty()) {
+            return null;
+        }
+        user = ValidatorControllers.validateUser(user);
+        User newUser = userService.update(user).get();
+        log.debug("Обновлен пользователь: {}", newUser);
+        return newUser;
     }
 
     @DeleteMapping
@@ -55,14 +62,18 @@ public class UserController {
 
     @GetMapping("/{userId}")
     // получение пользователя по id
-    public User findUserById(@PathVariable long userId) {
+    public Optional<User> findUserById(@PathVariable long userId) {
+        Optional<User> user = userService.findUserById(userId);
         log.debug("Получен пользователь с id = : {}", userId);
-        return userService.findUserById(userId);
+        return user;
     }
 
     @PutMapping("/{id}/friends/{friendId}")
     //  добавление в друзья
     public boolean addInFriends(@PathVariable long id, @PathVariable long friendId) {
+        if ((findUserById(id).isEmpty()) || (findUserById(friendId).isEmpty())) {
+            return false;
+        }
         log.debug("Пользователь c id = {} добавил в друзья пользователя с id = {}", id, friendId);
         return userService.addInFriends(id, friendId);
     }
@@ -70,6 +81,9 @@ public class UserController {
     @DeleteMapping("{id}/friends/{friendId}")
     //  удаление из друзей
     public boolean deleteFromFriends(@PathVariable long id, @PathVariable long friendId) {
+        if ((findUserById(id).isEmpty()) || (findUserById(friendId).isEmpty())) {
+            return false;
+        }
         log.debug("Пользователь c id = {} удалил из друзей пользователя с id = {}", id, friendId);
         return userService.deleteFromFriends(id, friendId);
     }
@@ -77,6 +91,9 @@ public class UserController {
     @GetMapping("/{id}/friends")
     // возвращаем список пользователей, являющихся его друзьями
     public List<User> findFriends(@PathVariable long id) {
+        if (findUserById(id).isEmpty()) {
+            return Collections.EMPTY_LIST;
+        }
         log.debug("Получен список пользователей, являющимися друзьями пользователя с id = {}", id);
         return userService.findFriends(id);
     }
@@ -84,6 +101,9 @@ public class UserController {
     @GetMapping("/{id}/friends/common/{otherId}")
     // список друзей, общих с другим пользователем
     public List<User> findMutualFriends(@PathVariable long id, @PathVariable long otherId) {
+        if ((findUserById(id).isEmpty()) || (findUserById(otherId).isEmpty())) {
+            return Collections.EMPTY_LIST;
+        }
         log.debug("Получен список друзей пользователя с id = {}, общих с пользователем с id = {}", id, otherId);
         return userService.findMutualFriends(id, otherId);
     }
